requirements.txt
pandas
pandas-gbq
pyyaml
google-cloud-bigquery
flask
flasgger
msal
Flask-JWT-Extended